webroot/requirements.txt
decorator==4.0.9
Django==1.9.2
django-allauth==0.24.1
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
ecdsa==0.13
httpie==0.9.3
ipython-genutils==0.1.0
Markdown==2.6.5
oauthlib==1.0.3
paramiko==1.16.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wheel==0.24.0
